// packages/astro/dist/assets/types.d.ts
export type ImageQuality = ImageQualityPreset | number;
export type ImageInputFormat = (typeof VALID_INPUT_FORMATS)[number];
export type ImageOutputFormat = (typeof VALID_OUTPUT_FORMATS)[number] | (string & {});
export type ImageLayout = 'constrained' | 'fixed' | 'full-width' | 'none';
export type ImageFit = 'fill' | 'contain' | 'cover' | 'none' | 'scale-down' | (string & {});
export type AssetsGlobalStaticImagesList = Map<
	string,
	{
		originalSrcPath: string | undefined;
		transforms: Map<
			string,
			{
				finalPath: string;
				transform: ImageTransform;
			}
		>;
	}
>;
// ...
export type ImageMetadata = {
	src: string;
	width: number;
	height: number;
	format: ImageInputFormat;
	orientation?: number;
	[isESMImport]?: true;
};
// ...
export type ImageTransform = {
	src: ImageMetadata | string;
	width?: number | undefined;
	widths?: number[] | undefined;
	densities?: (number | `${number}x`)[] | undefined;
	height?: number | undefined;
	quality?: ImageQuality | undefined;
	format?: ImageOutputFormat | undefined;
	fit?: ImageFit | undefined;
	position?: string | undefined;
	background?: string | undefined;
	[key: string]: any;
} & Astro.CustomImageProps;